+++ src/main/java/de/adorsys/keycloak/config/properties/ImportConfigProperties.java
@@ -25,6 +25,7 @@ import org.springframework.boot.context.properties.bind.DefaultValue;
 import org.springframework.validation.annotation.Validated;
 
 import java.util.Collection;
+import java.util.List;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
@@ -383,13 +384,18 @@ public class ImportConfigProperties {
         @NotNull
         private final ChecksumChangedOption checksumChanged;
 
+        @NotNull
+        private final Collection<String> userUpdateIgnoredProperties;
+
         public ImportBehaviorsProperties(boolean syncUserFederation, boolean removeDefaultRoleFromUser, boolean skipAttributesForFederatedUser,
-                                         boolean checksumWithCacheKey, ChecksumChangedOption checksumChanged) {
+                                         boolean checksumWithCacheKey, ChecksumChangedOption checksumChanged,
+                                         @DefaultValue("attributes") Collection<String> userUpdateIgnoredProperties) {
             this.syncUserFederation = syncUserFederation;
             this.removeDefaultRoleFromUser = removeDefaultRoleFromUser;
             this.skipAttributesForFederatedUser = skipAttributesForFederatedUser;
             this.checksumWithCacheKey = checksumWithCacheKey;
             this.checksumChanged = checksumChanged;
+            this.userUpdateIgnoredProperties = userUpdateIgnoredProperties == null ? List.of("attributes") : userUpdateIgnoredProperties;
         }
 
         public boolean isSyncUserFederation() {
@@ -410,6 +416,10 @@ public class ImportConfigProperties {
 
         public ChecksumChangedOption getChecksumChanged() {
             return checksumChanged;
+        }
+
+        public Collection<String> getUserUpdateIgnoredProperties() {
+            return userUpdateIgnoredProperties;
         }
 
         public enum ChecksumChangedOption {